refactor: move parseAssetIdForNavigation to TokenDetails as the Asset route's owned contract
Deletes the SocialLeaderboard tokenNavigation module (a copy of
PopularTokenRow's private helper) and extracts the original into
app/components/UI/TokenDetails/utils, imported by both call sites.
Removes the duplicated block flagged by the Sonar quality gate.
