+++ app/components/UI/TokenDetails/utils/parseAssetIdForNavigation.ts
@@ -6,6 +6,7 @@ import {
   parseCaipChainId,
 } from '@metamask/utils';
 
+// Zero address used for native EVM tokens (ETH, BNB, etc.)
 const NATIVE_TOKEN_ADDRESS = '0x0000000000000000000000000000000000000000';
 
 export interface AssetNavigationParams {
@@ -18,10 +19,14 @@ export interface AssetNavigationParams {
 }
 
 /**
- * Parses a CAIP-19 asset id into the params the wallet's `Asset` (token
- * details) route expects. Mirrors the local helper in
- * `Homepage/Sections/Tokens/components/PopularTokenRow.tsx` so a SocialLeaderboard
- * spot position can open the same token page — keep the two in sync.
+ * Parses a CAIP-19 asset id into the params the `Asset` (token details) route
+ * expects. Owned by TokenDetails as the destination's contract — any feature
+ * that opens the token page from a CAIP-19 asset id should use this instead of
+ * re-deriving the param shape locally.
+ *
+ * @param assetId - CAIP-19 asset id (e.g. "eip155:1/erc20:0x123...")
+ * @returns Parsed asset info with chainId, address, and type flags; empty
+ * strings and false flags when the asset id is not a valid CAIP-19 id.
  */
 export function parseAssetIdForNavigation(
   assetId: string,
@@ -51,10 +56,14 @@ export function parseAssetIdForNavigation(
     let address: string;
 
     if (isEvmChain) {
+      // EVM chains use hex chainId format for navigation
       chainId = `0x${parseInt(chainReference, 10).toString(16)}`;
+      // For native tokens (slip44), use zero address; for ERC20, use the contract address
       address = isNative ? NATIVE_TOKEN_ADDRESS : assetReference;
     } else {
+      // Non-EVM chains use CAIP-2 format for chainId
       chainId = `${namespace}:${chainReference}`;
+      // For non-EVM chains, address is the full CAIP-19 asset ID
       address = assetId;
     }
